Move `symbol_table_compiler` under `gem`
