--- lib/tapioca/compilers/gem/listeners.rb
@@ -1,16 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "tapioca/compilers/gem/listeners/base"
-require "tapioca/compilers/gem/listeners/dynamic_mixins"
-require "tapioca/compilers/gem/listeners/methods"
-require "tapioca/compilers/gem/listeners/mixins"
-require "tapioca/compilers/gem/listeners/remove_empty_payload_scopes"
-require "tapioca/compilers/gem/listeners/sorbet_enums"
-require "tapioca/compilers/gem/listeners/sorbet_helpers"
-require "tapioca/compilers/gem/listeners/sorbet_props"
-require "tapioca/compilers/gem/listeners/sorbet_required_ancestors"
-require "tapioca/compilers/gem/listeners/sorbet_signatures"
-require "tapioca/compilers/gem/listeners/sorbet_type_variables"
-require "tapioca/compilers/gem/listeners/subconstants"
-require "tapioca/compilers/gem/listeners/yard_doc"
