+++ lib/tapioca/gem/listeners.rb
@@ -1,0 +1,16 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "tapioca/gem/listeners/base"
+require "tapioca/gem/listeners/dynamic_mixins"
+require "tapioca/gem/listeners/methods"
+require "tapioca/gem/listeners/mixins"
+require "tapioca/gem/listeners/remove_empty_payload_scopes"
+require "tapioca/gem/listeners/sorbet_enums"
+require "tapioca/gem/listeners/sorbet_helpers"
+require "tapioca/gem/listeners/sorbet_props"
+require "tapioca/gem/listeners/sorbet_required_ancestors"
+require "tapioca/gem/listeners/sorbet_signatures"
+require "tapioca/gem/listeners/sorbet_type_variables"
+require "tapioca/gem/listeners/subconstants"
+require "tapioca/gem/listeners/yard_doc"
