+++ lib/tapioca/helpers/test/isolation.rb
@@ -100,7 +100,7 @@ module Tapioca
                   load_path_args << File.expand_path(p)
                 end
 
-                child = IO.popen([env, Gem.ruby, *load_path_args, $PROGRAM_NAME, *ORIG_ARGV, test_opts])
+                child = IO.popen([env, ::Gem.ruby, *load_path_args, $PROGRAM_NAME, *ORIG_ARGV, test_opts])
 
                 begin
                   Process.wait(child.pid)
