+++ lib/tapioca/internal.rb
@@ -18,7 +18,7 @@ require "tapioca/executor"
 require "tapioca/static/requires_compiler"
 require "tapioca/static/symbol_table_parser"
 require "tapioca/static/symbol_loader"
-require "tapioca/compilers/gem/events"
-require "tapioca/compilers/gem/listeners"
-require "tapioca/compilers/symbol_table_compiler"
+require "tapioca/gem/events"
+require "tapioca/gem/listeners"
+require "tapioca/gem/symbol_table_compiler"
 require "tapioca/dsl/pipeline"
